Fixed a syntax error
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12731 01821946-851b-0410-b7f4-f1477c70e57a

+++ network/network_sync.cpp
@@ -271,7 +271,7 @@ void NetworkSyncConnection (tNetworkSyncData *syncP)
 
 if (t < syncP->timeout)
 	return;
-syncP->timeout = t + 100 / Clamp (PacketsPerSec (), MIN_PPS, DEFAULT_PPS);
+syncP->timeout = t + 100 / Clamp (PacketsPerSec (), (short) MIN_PPS, (short) DEFAULT_PPS);
 #endif
 if (syncP->bExtraGameInfo) {
 	NetworkSendExtraGameInfo (&syncP->player [0]);
